requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==1.10.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
auto_gptq==0.7.1
autoawq==0.2.6
autoawq_kernels==0.0.7
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.8.0
bleach==6.1.0
blinker==1.8.2
Brotli==1.0.9
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-bullet==2.2.0
chroma-hnswlib==0.7.3
chroma-migrate==0.0.7
chromadb==0.5.3
click==8.1.7
clickhouse-connect==0.6.6
coloredlogs==15.0.1
cryptography==43.0.0
dataclasses-json==0.6.7
datasets==2.21.0
DateTime==5.5
decorator==5.1.1
deepdiff==8.0.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docx2txt==0.8
duckdb==0.7.1
EbookLib==0.17.1
einops==0.8.0
emoji==2.12.1
exceptiongroup==1.2.2
executing==2.1.0
extract-msg==0.23.1
faiss-cpu==1.8.0.post1
fastapi==0.112.3
fastjsonschema==2.20.0
filelock==3.13.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gekko==1.2.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.34.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
IMAPClient==2.1.0
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kubernetes==30.1.0
langchain==0.2.16
langchain-chroma==0.1.3
langchain-community==0.2.16
langchain-core==0.2.38
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.114
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.8.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
ollama==0.3.3
onnxruntime==1.19.2
openai==1.50.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
optimum==1.21.4
orderly-set==5.2.1
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pipreqs==0.5.0
platformdirs==4.3.6
posthog==3.6.3
prompt_toolkit==3.0.48
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-pptx==0.6.18
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
rapidfuzz==3.9.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.0
rouge==1.0.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==72.1.0
shellingham==1.5.4
six==1.12.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.4
streamlit==1.38.0
sudo==1.0.0
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
textract==1.6.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.3.1
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
triton==2.3.1
typer==0.12.5
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==1.5.1
unstructured==0.15.8
unstructured-client==0.25.5
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wheel==0.43.0
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.9.11
zipp==3.20.1
zope.interface==7.0.3
zstandard==0.23.0
